Updated to mingw64 version 4.9.3 and self compiled Qt libraries 5.4.2, available from http://flumes.iei.liu.se/hopsan/files/development_files/ Removed the _x64 extension from dependencies (to avoid confusion), (building both 32-bit and 64-bit) in the same directory is madness anyway. Use two different working directories instead.

+++ Dependencies/setupDiscount.bat
@@ -9,7 +9,6 @@ set filename=discount-2.1.8.zip
 set dirname=discount-2.1.8
 
 REM Automatic code begins here
-set dirname64=%dirname%_x64
 
 REM Unpack or checkout
 echo.
@@ -17,18 +16,10 @@ echo ======================
 echo Unpack Discount
 echo ======================
 echo Removing old directories (if they exist)
-rd /s/q %dirname%
-rd /s/q %dirname64%
+if exist %dirname% rd /s/q %dirname%
 mkdir %dirname%
 REM Unpack using tar
 ..\ThirdParty\7z\7z.exe x %filename% -y > nul
-
-REM Copy to 64-bit dir
-echo.
-echo ======================
-echo Copying to %dirname64%
-echo ======================
-robocopy /e /NFL /NDL /NJH /NJS /nc /ns /np  %dirname% %dirname64%
 
 REM Build
 echo.
@@ -37,7 +28,7 @@ echo Building 64-bit Discount
 echo ======================
 call setHopsanBuildPaths.bat 0.7.x x64
 
-cd %dirname64%
+cd %dirname%
 
 REM The first patch was taken from https://github.com/Alexpux/MINGW-packages/tree/master/mingw-w64-discount
 REM It was however modified to grep for Msys instead of MINGW (from uname -a)
